Fixes enums c#

+++ samples/client/petstore/python/swagger_client/apis/user_api.py
@@ -79,6 +79,7 @@ class UserApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/user'.replace('{format}', 'json')
         method = 'POST'
 
@@ -154,6 +155,7 @@ class UserApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/user/createWithArray'.replace('{format}', 'json')
         method = 'POST'
 
@@ -228,6 +230,7 @@ class UserApi(object):
                 )
             params[key] = val
         del params['kwargs']
+
 
         resource_path = '/user/createWithList'.replace('{format}', 'json')
         method = 'POST'
@@ -305,6 +308,7 @@ class UserApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/user/login'.replace('{format}', 'json')
         method = 'GET'
 
@@ -381,6 +385,7 @@ class UserApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/user/logout'.replace('{format}', 'json')
         method = 'GET'
 
@@ -440,9 +445,6 @@ class UserApi(object):
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'username' is set
-        if username is None:
-            raise ValueError("Missing the required parameter `username` when calling `get_user_by_name`")
 
         all_params = ['username']
         all_params.append('callback')
@@ -456,6 +458,10 @@ class UserApi(object):
                 )
             params[key] = val
         del params['kwargs']
+
+        # verify the required parameter 'username' is set
+        if ('username' not in params) or (params['username'] is None):
+            raise ValueError("Missing the required parameter `username` when calling `get_user_by_name`")
 
         resource_path = '/user/{username}'.replace('{format}', 'json')
         method = 'GET'
@@ -519,9 +525,6 @@ class UserApi(object):
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'username' is set
-        if username is None:
-            raise ValueError("Missing the required parameter `username` when calling `update_user`")
 
         all_params = ['username', 'body']
         all_params.append('callback')
@@ -535,6 +538,10 @@ class UserApi(object):
                 )
             params[key] = val
         del params['kwargs']
+
+        # verify the required parameter 'username' is set
+        if ('username' not in params) or (params['username'] is None):
+            raise ValueError("Missing the required parameter `username` when calling `update_user`")
 
         resource_path = '/user/{username}'.replace('{format}', 'json')
         method = 'PUT'
@@ -599,9 +606,6 @@ class UserApi(object):
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'username' is set
-        if username is None:
-            raise ValueError("Missing the required parameter `username` when calling `delete_user`")
 
         all_params = ['username']
         all_params.append('callback')
@@ -615,6 +619,10 @@ class UserApi(object):
                 )
             params[key] = val
         del params['kwargs']
+
+        # verify the required parameter 'username' is set
+        if ('username' not in params) or (params['username'] is None):
+            raise ValueError("Missing the required parameter `username` when calling `delete_user`")
 
         resource_path = '/user/{username}'.replace('{format}', 'json')
         method = 'DELETE'
